requirements.txt
livekit-agents
livekit-plugins-groq
livekit-plugins-silero   #used for voice activity detection
python-dotenv #loading our environment variable(env)
livekit-api
flask[async]
flask
flask-cors
uvicorn